% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_Finder.R
\name{Norm_Finder}
\alias{Norm_Finder}
\title{Find norms}
\usage{
Norm_Finder(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Dataframe of norms
}
\description{
Find norms
}
\examples{
sampledata <- data.frame(replicate(10,sample(0:100,1000,rep=TRUE)))
x <- sampledata$X1
Norm_Finder(x)
}
